chore(vip-650): rename VIP-636 to VIP-650
