+++ simulations/vip-650/bscmainnet.ts
@@ -3,7 +3,7 @@ import { ethers } from "hardhat";
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import vip636, { BSTOCK_LIQUIDATOR, UNITROLLER } from "../../vips/vip-636/bscmainnet";
+import vip650, { BSTOCK_LIQUIDATOR, UNITROLLER } from "../../vips/vip-650/bscmainnet";
 import FLASHLOAN_FACET_ABI from "./abi/FlashLoanFacet.json";
 
 const VUSDT = "0xfD5840Cd36d94D7229439859C0112a4185BC0255";
@@ -15,7 +15,7 @@ forking(FORK_BLOCK, async () => {
   const comptroller = new ethers.Contract(UNITROLLER, FLASHLOAN_FACET_ABI, ethers.provider);
   const vUSDT = new ethers.Contract(VUSDT, VUSDT_ABI, ethers.provider);
 
-  describe("VIP-636 Pre-VIP behavior", () => {
+  describe("VIP-650 Pre-VIP behavior", () => {
     it("BStockLiquidator is not yet whitelisted for flash loans", async () => {
       expect(await comptroller.authorizedFlashLoan(BSTOCK_LIQUIDATOR)).to.equal(false);
     });
@@ -25,13 +25,13 @@ forking(FORK_BLOCK, async () => {
     });
   });
 
-  testVip("VIP-636", await vip636(), {
+  testVip("VIP-650", await vip650(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [FLASHLOAN_FACET_ABI], ["IsAccountFlashLoanWhitelisted"], [1]);
     },
   });
 
-  describe("VIP-636 Post-VIP behavior", () => {
+  describe("VIP-650 Post-VIP behavior", () => {
     it("BStockLiquidator is whitelisted for flash loans", async () => {
       expect(await comptroller.authorizedFlashLoan(BSTOCK_LIQUIDATOR)).to.equal(true);
     });
